Remove remaining SERVER_URL references
- Resolve origin based on config
- Temporary keep config within legacy structure
