+++ .env.example
@@ -1,7 +1,8 @@
 # Server configuration
 HOSTNAME=127.0.0.1
-PROTOCOL=http
 PORT=3000
+PROTOCOL=http
+REVERSE_PROXY_PORT=
 
 # Database config
 DATABASE_URI=postgres://user:pass@hostname/database
