+++ server/shared/origin.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { hostname } = require('../config');
+const { hostname } = require('../../config/server');
 const logger = require('./logger');
 
 const isProduction = process.env.NODE_ENV === 'production';
